A little more refactoring.
