+++ src/tools/ZoomTool.js
@@ -1,7 +1,9 @@
 import external from '../externalModules.js';
 import BaseTool from '../base/BaseTool.js';
 import { clipToBox } from '../util/clip.js';
-import { correctShift, changeViewportScale } from '../util/zoom.js';
+import zoomUtils from '../util/zoom/index.js';
+
+const { correctShift, changeViewportScale } = zoomUtils;
 
 /**
  * @export @public @class
