+++ src/util/zoom/index.js
@@ -1,0 +1,7 @@
+import changeViewportScale from './changeViewportScale.js';
+import correctShift from './correctShift.js';
+
+export default {
+  changeViewportScale,
+  correctShift
+};
